bdi_api/s8/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
boto3
tqdm
psycopg2-binary
pydantic_settings>=2.0.0
pandas